+++ dart/test/http_client/sentry_http_client_test.dart
@@ -9,182 +9,159 @@ import 'package:test/test.dart';
 
 import '../mocks/mock_hub.dart';
 
+final requestUri = Uri.parse('https://example.com');
+
 void main() {
   group(SentryHttpClient, () {
+    late var fixture;
+
+    setUp(() {
+      fixture = Fixture();
+    });
+
     test('GET: happy path', () async {
-      final mockHub = MockHub();
+      final sut =
+          fixture.getSut(fixture.getClient(statusCode: 200, reason: 'OK'));
 
-      final mockClient = MockClient((request) async {
-        expect(request.url, Uri.parse('https://example.com'));
-        return Response('', 200, reasonPhrase: 'OK');
-      });
-
-      final client = SentryHttpClient(client: mockClient, hub: mockHub);
-
-      final response = await client.get(Uri.parse('https://example.com'));
+      final response = await sut.get(requestUri);
       expect(response.statusCode, 200);
 
-      expect(mockHub.addBreadcrumbCalls.length, 1);
-      final breadcrumb = mockHub.addBreadcrumbCalls.first.crumb;
+      expect(fixture.hub.addBreadcrumbCalls.length, 1);
+      final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data, <String, dynamic>{
-        'url': 'https://example.com',
-        'method': 'GET',
-        'status_code': 200,
-        'reason': 'OK',
-      });
+      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['method'], 'GET');
+      expect(breadcrumb.data?['status_code'], 200);
+      expect(breadcrumb.data?['reason'], 'OK');
+      expect(breadcrumb.data?['duration'], isNotNull);
     });
 
     test('GET: happy path for 404', () async {
-      final mockHub = MockHub();
+      final sut = fixture
+          .getSut(fixture.getClient(statusCode: 404, reason: 'NOT FOUND'));
 
-      final mockClient = MockClient((request) async {
-        expect(request.url, Uri.parse('https://example.com'));
-        return Response('', 404, reasonPhrase: 'NOT FOUND');
-      });
+      final response = await sut.get(requestUri);
 
-      final client = SentryHttpClient(client: mockClient, hub: mockHub);
-
-      final response = await client.get(Uri.parse('https://example.com'));
       expect(response.statusCode, 404);
 
-      expect(mockHub.addBreadcrumbCalls.length, 1);
-      final breadcrumb = mockHub.addBreadcrumbCalls.first.crumb;
+      expect(fixture.hub.addBreadcrumbCalls.length, 1);
+      final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data, <String, dynamic>{
-        'url': 'https://example.com',
-        'method': 'GET',
-        'status_code': 404,
-        'reason': 'NOT FOUND',
-      });
+      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['method'], 'GET');
+      expect(breadcrumb.data?['status_code'], 404);
+      expect(breadcrumb.data?['reason'], 'NOT FOUND');
+      expect(breadcrumb.data?['duration'], isNotNull);
     });
 
     test('POST: happy path', () async {
-      final mockHub = MockHub();
+      final sut = fixture.getSut(fixture.getClient(statusCode: 200));
 
-      final mockClient = MockClient((request) async {
-        expect(request.url, Uri.parse('https://example.com'));
-        return Response('', 200);
-      });
-
-      final client = SentryHttpClient(client: mockClient, hub: mockHub);
-
-      final response = await client.post(Uri.parse('https://example.com'));
+      final response = await sut.post(requestUri);
       expect(response.statusCode, 200);
 
-      expect(mockHub.addBreadcrumbCalls.length, 1);
-      final breadcrumb = mockHub.addBreadcrumbCalls.first.crumb;
+      expect(fixture.hub.addBreadcrumbCalls.length, 1);
+      final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data, <String, dynamic>{
-        'url': 'https://example.com',
-        'method': 'POST',
-        'status_code': 200,
-      });
+      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['method'], 'POST');
+      expect(breadcrumb.data?['status_code'], 200);
+      expect(breadcrumb.data?['duration'], isNotNull);
     });
 
     test('PUT: happy path', () async {
-      final mockHub = MockHub();
+      final sut = fixture.getSut(fixture.getClient(statusCode: 200));
 
-      final mockClient = MockClient((request) async {
-        expect(request.url, Uri.parse('https://example.com'));
-        return Response('', 200);
-      });
-
-      final client = SentryHttpClient(client: mockClient, hub: mockHub);
-
-      final response = await client.put(Uri.parse('https://example.com'));
+      final response = await sut.put(requestUri);
       expect(response.statusCode, 200);
 
-      expect(mockHub.addBreadcrumbCalls.length, 1);
-      final breadcrumb = mockHub.addBreadcrumbCalls.first.crumb;
+      expect(fixture.hub.addBreadcrumbCalls.length, 1);
+      final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data, <String, dynamic>{
-        'url': 'https://example.com',
-        'method': 'PUT',
-        'status_code': 200,
-      });
+      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['method'], 'PUT');
+      expect(breadcrumb.data?['status_code'], 200);
+      expect(breadcrumb.data?['duration'], isNotNull);
     });
 
     test('DELETE: happy path', () async {
-      final mockHub = MockHub();
+      final sut = fixture.getSut(fixture.getClient(statusCode: 200));
 
-      final mockClient = MockClient((request) async {
-        expect(request.url, Uri.parse('https://example.com'));
-        return Response('', 200, reasonPhrase: 'OK');
-      });
-
-      final client = SentryHttpClient(client: mockClient, hub: mockHub);
-
-      final response = await client.delete(Uri.parse('https://example.com'));
+      final response = await sut.delete(requestUri);
       expect(response.statusCode, 200);
 
-      expect(mockHub.addBreadcrumbCalls.length, 1);
-      final breadcrumb = mockHub.addBreadcrumbCalls.first.crumb;
+      expect(fixture.hub.addBreadcrumbCalls.length, 1);
+      final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data, <String, dynamic>{
-        'url': 'https://example.com',
-        'method': 'DELETE',
-        'status_code': 200,
-        'reason': 'OK',
-      });
+      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['method'], 'DELETE');
+      expect(breadcrumb.data?['status_code'], 200);
+      expect(breadcrumb.data?['duration'], isNotNull);
     });
 
     /// Tests, that in case an exception gets thrown, that
-    ///   - no breadcrumb gets added
-    ///   - no exception gets reported by Sentry, in case the user wants to
-    ///     handle the exception
-    test('no breadcrumb for ClientException', () async {
-      final url = Uri.parse('https://example.com');
-
-      final mockHub = MockHub();
-
-      final mockClient = MockClient((request) async {
-        expect(request.url, url);
-        throw ClientException('test', url);
-      });
-
-      final client = SentryHttpClient(client: mockClient, hub: mockHub);
+    /// no exception gets reported by Sentry, in case the user wants to
+    /// handle the exception
+    test('no captureException for ClientException', () async {
+      final sut = fixture.getSut(MockClient((request) async {
+        expect(request.url, requestUri);
+        throw ClientException('test', requestUri);
+      }));
 
       try {
-        await client.get(Uri.parse('https://example.com'));
+        await sut.get(requestUri);
         fail('Method did not throw');
       } on ClientException catch (e) {
         expect(e.message, 'test');
-        expect(e.uri, url);
+        expect(e.uri, requestUri);
       }
 
-      expect(mockHub.addBreadcrumbCalls.length, 0);
-      expect(mockHub.captureExceptionCalls.length, 0);
+      expect(fixture.hub.captureExceptionCalls.length, 0);
     });
 
     /// SocketException are only a thing on dart:io platforms.
     /// otherwise this is equal to the test above
-    test('no breadcrumb for SocketException', () async {
-      final url = Uri.parse('https://example.com');
-
-      final mockHub = MockHub();
-
-      final mockClient = MockClient((request) async {
-        expect(request.url, url);
+    test('no captureException for SocketException', () async {
+      final sut = fixture.getSut(MockClient((request) async {
+        expect(request.url, requestUri);
         throw SocketException('test');
-      });
-
-      final client = SentryHttpClient(client: mockClient, hub: mockHub);
+      }));
 
       try {
-        await client.get(Uri.parse('https://example.com'));
+        await sut.get(requestUri);
         fail('Method did not throw');
       } on SocketException catch (e) {
         expect(e.message, 'test');
       }
 
-      expect(mockHub.addBreadcrumbCalls.length, 0);
-      expect(mockHub.captureExceptionCalls.length, 0);
+      expect(fixture.hub.captureExceptionCalls.length, 0);
+    });
+
+    test('breadcrumb gets added when an exception gets thrown', () async {
+      final sut = fixture.getSut(MockClient((request) async {
+        expect(request.url, requestUri);
+        throw Exception('foo bar');
+      }));
+
+      try {
+        await sut.get(requestUri);
+        fail('Method did not throw');
+      } on Exception catch (_) {}
+
+      expect(fixture.hub.addBreadcrumbCalls.length, 1);
+
+      final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
+
+      expect(breadcrumb.type, 'http');
+      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['method'], 'GET');
+      expect(breadcrumb.level, SentryLevel.error);
+      expect(breadcrumb.data?['duration'], isNotNull);
     });
 
     test('close does get called for user defined client', () async {
@@ -199,7 +176,41 @@ void main() {
       expect(mockHub.captureExceptionCalls.length, 0);
       verify(mockClient.close());
     });
+
+    test('Breadcrumb has correct duration', () async {
+      final sut = fixture.getSut(MockClient((request) async {
+        expect(request.url, requestUri);
+        await Future.delayed(Duration(seconds: 1));
+        return Response('', 200, reasonPhrase: 'OK');
+      }));
+
+      final response = await sut.get(requestUri);
+      expect(response.statusCode, 200);
+
+      expect(fixture.hub.addBreadcrumbCalls.length, 1);
+      final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
+
+      var durationString = breadcrumb.data!['duration']! as String;
+      // we don't check for anything below a second
+      expect(durationString.startsWith('0:00:01'), true);
+    });
   });
 }
 
 class CloseableMockClient extends Mock implements BaseClient {}
+
+class Fixture {
+  SentryHttpClient getSut([MockClient? client]) {
+    final mc = client ?? getClient();
+    return SentryHttpClient(client: mc, hub: hub);
+  }
+
+  late MockHub hub = MockHub();
+
+  MockClient getClient({int statusCode = 200, String? reason}) {
+    return MockClient((request) async {
+      expect(request.url, requestUri);
+      return Response('', statusCode, reasonPhrase: reason);
+    });
+  }
+}