+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Feat: Support for Windows (#407)
 * Fix: Mark `Sentry.currentHub` as deprecated (#406)
 * Fix: Use name from pubspec.yaml for release if package id is not available (#411)
+* Feat: `SentryHttpClient` tracks the duration which a request takes and logs failed requests (#414)
 
 # 5.0.0